script/requirements.txt
beautifulsoup4
PyYAML
requests
rich
typer
